Same page refresh update to Home page
Add extra route to react router in App, update to Home to accept page
parameters when navigating to new “popular” route and update
PageNavigator component to ensure the user is navigated to “popular”
route when all nav buttons are clicked.

Update to MoviePreviewInfo to navigate the user to MovieInfo when the
backdrop image is now clicked, add a method (that enables scrolling) to
be invoked every time the component is unmounted, and minor styling
changes.